refactor #29: 분석 로직 analysis 패키지로 이동
